+++ TO2/AST/FieldAssign.cs
@@ -114,6 +114,8 @@ namespace KontrolSystem.TO2.AST {
             }
 
             if (op == Operator.Assign) {
+                expression.Prepare(context);
+                
                 if (fieldAccess.RequiresPtr) target.EmitPtr(context);
                 else target.EmitCode(context, false);