+++ CHANGES.md
@@ -14,6 +14,7 @@
 * Add improved launch sequence (thanks to RobertoBiundo)
 * Extend the resource transfer API to support transfer of specific resources (and specific number of units)
 * Fix NPE in type checking
+* Fix broken error reporting (issue #78)
 
 ## 0.3.3.2 -> 0.3.3.3